Leitura do .class

+++ header/Dados.hpp
@@ -4,6 +4,6 @@
 typedef unsigned char u1;
 typedef unsigned short u2;
 typedef unsigned int u4;
-typedef unsigned long u8;
+typedef unsigned long long u8;
 
 #endif
